Evergreen change to leverage existing arg validation utility methods
