+++ src/test/shared/sam/cli/samCliTestUtils.ts
@@ -5,6 +5,7 @@
 
 'use strict'
 
+import * as assert from 'assert'
 import { SpawnOptions } from 'child_process'
 import { SamCliProcessInvoker } from '../../../../shared/sam/cli/samCliInvokerUtils'
 import { ChildProcessResult } from '../../../../shared/utilities/childProcess'
@@ -25,4 +26,37 @@ export class MockSamCliProcessInvoker implements SamCliProcessInvoker {
             exitCode: 0
         } as any as ChildProcessResult
     }
+}
+
+export function assertArgsContainArgument(
+    args: any[],
+    argOfInterest: string,
+    expectedArgValue: string
+) {
+    const argPos = args.indexOf(argOfInterest)
+    assert.notStrictEqual(argPos, -1, `Expected arg ${argOfInterest} was not found`)
+    assert.ok(args.length >= argPos + 2, `Args does not contain a value for ${argOfInterest}`)
+    assert.strictEqual(args[argPos + 1], expectedArgValue, `Arg ${argOfInterest} did not have expected value`)
+}
+
+export function assertArgIsPresent(
+    args: any[],
+    argOfInterest: string,
+) {
+    assert.notStrictEqual(
+        args.indexOf(argOfInterest),
+        -1,
+        `Expected ${argOfInterest} arg`
+    )
+}
+
+export function assertArgNotPresent(
+    args: any[],
+    argOfInterest: string,
+) {
+    assert.strictEqual(
+        args.indexOf(argOfInterest),
+        -1,
+        `Did not expect ${argOfInterest} arg`
+    )
 }
